Move PartiQLValueComparator to partiql-types and make a public fn
